Respond to PR feedback add additional tests
